refactor(Test): move mocks into __mocks__ directory
