+++ tests/__mocks__/bitcoin-exchange-client.mock.js
