+++ tests/__mocks__/spender_mock.js
